DROID-3429 refactoring
