--- feature-object-type/src/main/java/com/anytypeio/anytype/feature_object_type/properties/edit/ui/PropertyLimitObjectTypeScreen.kt
@@ -1,9 +1,0 @@
-package com.anytypeio.anytype.feature_object_type.properties.edit.ui
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun PropertyLimitObjectTypeScreen() {
-
-
-}
