+++ feature-properties/src/main/java/com/anytypeio/anytype/feature_properties/EditSpacePropertiesViewModel.kt
@@ -1,4 +1,4 @@
-package com.anytypeio.anytype.feature_object_type.properties
+package com.anytypeio.anytype.feature_properties
 
 class EditSpacePropertiesViewModel {
     //todo: implement later
