+++ feature-properties/src/main/java/com/anytypeio/anytype/feature_properties/VMFactory.kt
@@ -1,4 +1,4 @@
-package com.anytypeio.anytype.feature_object_type.properties
+package com.anytypeio.anytype.feature_properties
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -8,7 +8,7 @@ import com.anytypeio.anytype.domain.objects.StoreOfRelations
 import com.anytypeio.anytype.domain.primitives.SetObjectTypeRecommendedFields
 import com.anytypeio.anytype.domain.relations.CreateRelation
 import com.anytypeio.anytype.domain.resources.StringResourceProvider
-import com.anytypeio.anytype.feature_object_type.properties.add.AddPropertyVmParams
+import com.anytypeio.anytype.feature_properties.add.AddPropertyVmParams
 import javax.inject.Inject
 
 class AddPropertyVmFactory @Inject constructor(
