+++ feature-properties/src/main/java/com/anytypeio/anytype/feature_properties/edit/UiEditPropertyState.kt
@@ -1,10 +1,9 @@
-package com.anytypeio.anytype.feature_object_type.properties.edit
+package com.anytypeio.anytype.feature_properties.edit
 
 import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.core_models.Key
 import com.anytypeio.anytype.core_models.RelationFormat
-import com.anytypeio.anytype.feature_object_type.fields.UiFieldObjectItem
-import com.anytypeio.anytype.presentation.objects.UiObjectsListItem
+import com.anytypeio.anytype.presentation.objects.ObjectIcon
 
 sealed class UiEditPropertyState {
     data object Hidden : UiEditPropertyState()
@@ -17,7 +16,7 @@ sealed class UiEditPropertyState {
             val formatName: String,
             val formatIcon: Int?,
             val format: RelationFormat,
-            val limitObjectTypes: List<UiFieldObjectItem> = emptyList()
+            val limitObjectTypes: List<UiPropertyLimitTypeItem> = emptyList()
         ) : Visible()
 
         data class New(
@@ -25,7 +24,7 @@ sealed class UiEditPropertyState {
             val formatName: String,
             val formatIcon: Int?,
             val format: RelationFormat,
-            val limitObjectTypes: List<UiFieldObjectItem> = emptyList()
+            val limitObjectTypes: List<UiPropertyLimitTypeItem> = emptyList()
         ) : Visible()
 
         data class View(
@@ -35,17 +34,11 @@ sealed class UiEditPropertyState {
             val formatName: String,
             val formatIcon: Int?,
             val format: RelationFormat,
-            val limitObjectTypes: List<UiFieldObjectItem> = emptyList()
+            val limitObjectTypes: List<UiPropertyLimitTypeItem> = emptyList()
         ) : Visible()
     }
 }
 
-sealed class UiPropertyLimitObjectTypesState {
-    data class Edit(
-        val limitObjectTypes: List<UiObjectsListItem>
-    ) : UiPropertyLimitObjectTypesState()
-
-    data class View(
-        val limitObjectTypes: List<UiObjectsListItem>
-    ) : UiPropertyLimitObjectTypesState()
-}
+data class UiPropertyLimitTypeItem(
+    val id: Id, val key: Key, val title: String, val icon: ObjectIcon
+)
