+++ feature-properties/src/main/java/com/anytypeio/anytype/feature_properties/edit/ui/EditPropertyScreen.kt
@@ -1,4 +1,4 @@
-package com.anytypeio.anytype.feature_object_type.properties.edit.ui
+package com.anytypeio.anytype.feature_properties.edit.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
@@ -30,7 +30,7 @@ import com.anytypeio.anytype.core_ui.foundation.Divider
 import com.anytypeio.anytype.core_ui.foundation.noRippleThrottledClickable
 import com.anytypeio.anytype.core_ui.views.ButtonPrimary
 import com.anytypeio.anytype.core_ui.views.ButtonSize
-import com.anytypeio.anytype.feature_object_type.properties.edit.UiEditPropertyState
+import com.anytypeio.anytype.feature_properties.edit.UiEditPropertyState
 
 @Composable
 fun PropertyEditScreen(
